requirements.txt
Flask==3.0.3
google-cloud-datastore==2.19.0
